+++ tests/test_lab_dev/test_states.py
@@ -21,12 +21,14 @@ import os
 import numpy as np
 import pytest
 
-from mrmustard import math
+from mrmustard import math, settings
+from mrmustard.math.parameters import Constant, Variable
 from mrmustard.physics.representations import Bargmann
 from mrmustard.physics.fock import fock_state
 from mrmustard.physics.gaussian import vacuum_cov, vacuum_means, squeezed_vacuum_cov
 from mrmustard.physics.triples import coherent_state_Abc, thermal_state_Abc
 from mrmustard.lab_dev.circuit_components import CircuitComponent
+from mrmustard.lab_dev.circuit_components_utils import TraceOut
 from mrmustard.lab_dev.states import (
     Coherent,
     DisplacedSqueezed,
@@ -39,6 +41,9 @@ from mrmustard.lab_dev.states import (
 )
 from mrmustard.lab_dev.transformations import Attenuator, Dgate, Sgate
 from mrmustard.lab_dev.wires import Wires
+
+# original settings
+autocutoff_max0 = settings.AUTOCUTOFF_MAX_CUTOFF
 
 
 class TestKet:
@@ -99,8 +104,10 @@ class TestKet:
 
     @pytest.mark.parametrize("modes", [[0], [0, 1], [3, 19, 2]])
     def test_to_from_phase_space(self, modes):
-        with pytest.raises(NotImplementedError):
-            Coherent(modes, x=1, y=2).phase_space()
+        cov, means, coeff = Coherent([0], x=1, y=2).phase_space(s=0)
+        assert math.allclose(coeff[0], 1.0)
+        assert math.allclose(cov[0], np.eye(2))
+        assert math.allclose(means[0], np.array([2.0, 4.0]))
 
         n_modes = len(modes)
 
@@ -153,6 +160,107 @@ class TestKet:
         assert dm.representation == (ket @ ket.adjoint).representation
         assert dm.wires == (ket @ ket.adjoint).wires
 
+    def test_expectation_bargmann(self):
+        ket = Coherent([0, 1], x=1, y=[2, 3])
+
+        assert math.allclose(ket.expectation(ket), [1])
+
+        k0 = Coherent([0], x=1, y=2)
+        k1 = Coherent([1], x=1, y=3)
+        k01 = Coherent([0, 1], x=1, y=[2, 3])
+
+        res_k0 = ((ket @ k0.dual) >> TraceOut([1])).representation.c
+        res_k1 = ((ket @ k1.dual) >> TraceOut([0])).representation.c
+        res_k01 = (ket @ k01.dual).representation.c
+
+        assert math.allclose(ket.expectation(k0), res_k0)
+        assert math.allclose(ket.expectation(k1), res_k1)
+        assert math.allclose(ket.expectation(k01), res_k01)
+
+        dm0 = Coherent([0], x=1, y=2).dm()
+        dm1 = Coherent([1], x=1, y=3).dm()
+        dm01 = Coherent([0, 1], x=1, y=[2, 3]).dm()
+
+        res_dm0 = ((ket @ ket.adjoint @ dm0.dual) >> TraceOut([1])).representation.c
+        res_dm1 = ((ket @ ket.adjoint @ dm1.dual) >> TraceOut([0])).representation.c
+        res_dm01 = (ket @ ket.adjoint @ dm01.dual).representation.c
+
+        assert math.allclose(ket.expectation(dm0), res_dm0)
+        assert math.allclose(ket.expectation(dm1), res_dm1)
+        assert math.allclose(ket.expectation(dm01), res_dm01)
+
+        u0 = Dgate([1], x=0.1)
+        u1 = Dgate([0], x=0.2)
+        u01 = Dgate([0, 1], x=[0.3, 0.4])
+
+        res_u0 = (ket @ u0 @ ket.dual).representation.c
+        res_u1 = (ket @ u1 @ ket.dual).representation.c
+        res_u01 = (ket @ u01 @ ket.dual).representation.c
+
+        assert math.allclose(ket.expectation(u0), res_u0)
+        assert math.allclose(ket.expectation(u1), res_u1)
+        assert math.allclose(ket.expectation(u01), res_u01)
+
+    def test_expectation_fock(self):
+        settings.AUTOCUTOFF_MAX_CUTOFF = 10
+
+        ket = Coherent([0, 1], x=1, y=[2, 3]).to_fock_component()
+
+        assert math.allclose(ket.expectation(ket), (ket @ ket.dual).representation.array ** 2)
+
+        k0 = Coherent([0], x=1, y=2)
+        k1 = Coherent([1], x=1, y=3)
+        k01 = Coherent([0, 1], x=1, y=[2, 3])
+
+        res_k0 = ((ket @ k0.dual) >> TraceOut([1])).representation.array
+        res_k1 = ((ket @ k1.dual) >> TraceOut([0])).representation.array
+        res_k01 = (ket @ k01.dual).representation.array ** 2
+
+        assert math.allclose(ket.expectation(k0), res_k0)
+        assert math.allclose(ket.expectation(k1), res_k1)
+        assert math.allclose(ket.expectation(k01), res_k01)
+
+        dm0 = Coherent([0], x=1, y=2).dm()
+        dm1 = Coherent([1], x=1, y=3).dm()
+        dm01 = Coherent([0, 1], x=1, y=[2, 3]).dm()
+
+        res_dm0 = ((ket @ ket.adjoint @ dm0.dual) >> TraceOut([1])).representation.array
+        res_dm1 = ((ket @ ket.adjoint @ dm1.dual) >> TraceOut([0])).representation.array
+        res_dm01 = (ket @ ket.adjoint @ dm01.dual).representation.array
+
+        assert math.allclose(ket.expectation(dm0), res_dm0)
+        assert math.allclose(ket.expectation(dm1), res_dm1)
+        assert math.allclose(ket.expectation(dm01), res_dm01)
+
+        u0 = Dgate([1], x=0.1)
+        u1 = Dgate([0], x=0.2)
+        u01 = Dgate([0, 1], x=[0.3, 0.4])
+
+        res_u0 = (ket @ u0 @ ket.dual).representation.array
+        res_u1 = (ket @ u1 @ ket.dual).representation.array
+        res_u01 = (ket @ u01 @ ket.dual).representation.array
+
+        assert math.allclose(ket.expectation(u0), res_u0)
+        assert math.allclose(ket.expectation(u1), res_u1)
+        assert math.allclose(ket.expectation(u01), res_u01)
+
+        settings.AUTOCUTOFF_MAX_CUTOFF = autocutoff_max0
+
+    def test_expectation_error(self):
+        ket = Coherent([0, 1], x=1, y=[2, 3])
+
+        op1 = Attenuator([0])
+        with pytest.raises(ValueError, match="Cannot calculate the expectation value"):
+            ket.expectation(op1)
+
+        op2 = CircuitComponent("", None, modes_in_ket=[0], modes_out_ket=[1])
+        with pytest.raises(ValueError, match="different modes"):
+            ket.expectation(op2)
+
+        op3 = Dgate([2])
+        with pytest.raises(ValueError, match="Expected an operator defined on"):
+            ket.expectation(op3)
+
     def test_rshift(self):
         ket = Coherent([0, 1], 1)
         unitary = Dgate([0], 1)
@@ -175,6 +283,39 @@ class TestKet:
         # measurements
         assert isinstance(ket >> Coherent([0], 1).dual, Ket)
         assert isinstance(ket >> Coherent([0], 1).dm().dual, DM)
+
+    @pytest.mark.parametrize("modes", [[3, 30, 98]])
+    @pytest.mark.parametrize("m", [[3], [30], [98], [3, 98]])
+    def test_get_item(self, modes, m):
+        ket = Vacuum(modes) >> Dgate(modes, x=[0, 1, 2])
+        dm = ket.dm()
+
+        assert ket[m] == dm[m]
+
+    @pytest.mark.parametrize("modes", [[3, 30, 98]])
+    @pytest.mark.parametrize("m", [[3], [30], [98], [3, 98]])
+    def test_get_item_builtin_kets(self, modes, m):
+        idx = [modes.index(s) for s in m]
+
+        x = math.asnumpy([0, 1, 2])
+        s = DisplacedSqueezed(modes, x=x, y=3, y_trainable=True, y_bounds=(0, 6))
+
+        si = s[m]
+        assert isinstance(si, DisplacedSqueezed)
+        assert si == DisplacedSqueezed(m, x=x[idx], y=3, y_trainable=True, y_bounds=(0, 6))
+
+        assert isinstance(si.x, Constant)
+        assert math.allclose(si.x.value, x[idx])
+
+        assert isinstance(si.y, Variable)
+        assert si.y.value == s.y.value
+        assert si.y.bounds == s.y.bounds
+
+        assert isinstance(si.r, Constant)
+        assert si.r.value == s.r.value
+
+        assert isinstance(si.phi, Constant)
+        assert si.phi.value == s.phi.value
 
 
 class TestDM:
@@ -227,10 +368,11 @@ class TestDM:
         assert state_in_fock == state_out
 
     def test_to_from_phase_space(self):
-        state0 = Coherent([0], x=1, y=2) >> Attenuator([0], 0.8)
-
-        with pytest.raises(NotImplementedError):
-            state0.phase_space()
+        state0 = Coherent([0], x=1, y=2) >> Attenuator([0], 1.0)
+        cov, means, coeff = state0.phase_space(s=0)  # batch = 1
+        assert coeff[0] == 1.0
+        assert math.allclose(cov[0], np.eye(2))
+        assert math.allclose(means[0], np.array([2.0, 4.0]))
 
         cov = vacuum_cov(1)
         means = [1.78885438, 3.57770876]
@@ -259,7 +401,7 @@ class TestDM:
 
         state2 = Coherent([0], x=1).dm() / 3 + 2 * Coherent([0], x=-1).dm() / 3
         assert state2.probability == 1
-        assert state2.to_fock_component(20).probability == 1
+        assert math.allclose(state2.to_fock_component(20).probability, 1)
 
         state3 = Number([0], n=1, cutoffs=2).dm() / 2 + Number([0], n=2).dm() / 2
         assert math.allclose(state3.probability, 1)
@@ -268,6 +410,105 @@ class TestDM:
         state = Coherent([0], 1, 2).dm()
         assert math.allclose(state.purity, 1)
         assert state.is_pure
+
+    def test_expectation_bargmann(self):
+        ket = Coherent([0, 1], x=1, y=[2, 3])
+        dm = ket.dm()
+
+        k0 = Coherent([0], x=1, y=2)
+        k1 = Coherent([1], x=1, y=3)
+        k01 = Coherent([0, 1], x=1, y=[2, 3])
+
+        res_k0 = ((dm @ k0.dual @ k0.dual.adjoint) >> TraceOut([1])).representation.c
+        res_k1 = ((dm @ k1.dual @ k1.dual.adjoint) >> TraceOut([0])).representation.c
+        res_k01 = (dm @ k01.dual @ k01.dual.adjoint).representation.c
+
+        assert math.allclose(dm.expectation(k0), res_k0)
+        assert math.allclose(dm.expectation(k1), res_k1)
+        assert math.allclose(dm.expectation(k01), res_k01)
+
+        dm0 = Coherent([0], x=1, y=2).dm()
+        dm1 = Coherent([1], x=1, y=3).dm()
+        dm01 = Coherent([0, 1], x=1, y=[2, 3]).dm()
+
+        res_dm0 = ((dm @ dm0.dual) >> TraceOut([1])).representation.c
+        res_dm1 = ((dm @ dm1.dual) >> TraceOut([0])).representation.c
+        res_dm01 = (dm @ dm01.dual).representation.c
+
+        assert math.allclose(dm.expectation(dm0), res_dm0)
+        assert math.allclose(dm.expectation(dm1), res_dm1)
+        assert math.allclose(dm.expectation(dm01), res_dm01)
+
+        u0 = Dgate([0], x=0.1)
+        u1 = Dgate([1], x=0.2)
+        u01 = Dgate([0, 1], x=[0.3, 0.4])
+
+        res_u0 = ((dm @ u0) >> TraceOut([0, 1])).representation.c
+        res_u1 = ((dm @ u1) >> TraceOut([0, 1])).representation.c
+        res_u01 = ((dm @ u01) >> TraceOut([0, 1])).representation.c
+
+        assert math.allclose(dm.expectation(u0), res_u0)
+        assert math.allclose(dm.expectation(u1), res_u1)
+        assert math.allclose(dm.expectation(u01), res_u01)
+
+    def test_expectation_fock(self):
+        settings.AUTOCUTOFF_MAX_CUTOFF = 10
+
+        ket = Coherent([0, 1], x=1, y=[2, 3]).to_fock_component()
+        dm = ket.dm()
+
+        k0 = Coherent([0], x=1, y=2)
+        k1 = Coherent([1], x=1, y=3)
+        k01 = Coherent([0, 1], x=1, y=[2, 3])
+
+        res_k0 = ((dm @ k0.dual @ k0.dual.adjoint) >> TraceOut([1])).representation.array
+        res_k1 = ((dm @ k1.dual @ k1.dual.adjoint) >> TraceOut([0])).representation.array
+        res_k01 = (dm @ k01.dual @ k01.dual.adjoint).representation.array
+
+        assert math.allclose(dm.expectation(k0), res_k0)
+        assert math.allclose(dm.expectation(k1), res_k1)
+        assert math.allclose(dm.expectation(k01), res_k01)
+
+        dm0 = Coherent([0], x=1, y=2).dm()
+        dm1 = Coherent([1], x=1, y=3).dm()
+        dm01 = Coherent([0, 1], x=1, y=[2, 3]).dm()
+
+        res_dm0 = ((dm @ dm0.dual) >> TraceOut([1])).representation.array
+        res_dm1 = ((dm @ dm1.dual) >> TraceOut([0])).representation.array
+        res_dm01 = (dm @ dm01.dual).representation.array
+
+        assert math.allclose(dm.expectation(dm0), res_dm0)
+        assert math.allclose(dm.expectation(dm1), res_dm1)
+        assert math.allclose(dm.expectation(dm01), res_dm01)
+
+        u0 = Dgate([0], x=0.1)
+        u1 = Dgate([1], x=0.2)
+        u01 = Dgate([0, 1], x=[0.3, 0.4])
+
+        res_u0 = ((dm @ u0) >> TraceOut([0, 1])).representation.array
+        res_u1 = ((dm @ u1) >> TraceOut([0, 1])).representation.array
+        res_u01 = ((dm @ u01) >> TraceOut([0, 1])).representation.array
+
+        assert math.allclose(dm.expectation(u0), res_u0)
+        assert math.allclose(dm.expectation(u1), res_u1)
+        assert math.allclose(dm.expectation(u01), res_u01)
+
+        settings.AUTOCUTOFF_MAX_CUTOFF = autocutoff_max0
+
+    def test_expectation_error(self):
+        dm = Coherent([0, 1], x=1, y=[2, 3]).dm()
+
+        op1 = Attenuator([0])
+        with pytest.raises(ValueError, match="Cannot calculate the expectation value"):
+            dm.expectation(op1)
+
+        op2 = CircuitComponent("", None, modes_in_ket=[0], modes_out_ket=[1])
+        with pytest.raises(ValueError, match="different modes"):
+            dm.expectation(op2)
+
+        op3 = Dgate([2])
+        with pytest.raises(ValueError, match="Expected an operator defined on"):
+            dm.expectation(op3)
 
     def test_rshift(self):
         ket = Coherent([0, 1], 1)